refactor(analytics): extract shared helpers, remove redundant event scanning
- Extract formatCompact, formatDuration, formatDurationMs, barForPercent to
  shared frontend/src/lib/utils.ts (was duplicated in 2 components)
- Use stored rollup.toolUseCounts for /sessions/:id instead of re-scanning
  raw events
- Simplify rate-limit snapshot interval (5m directly, not 60s with guard)
- Single-pass partition in flushEventBuffer instead of two filter scans
- Use path.basename() instead of split/pop for repo name fallback
- Add sessionId validation to GET /sessions/:id
- Add Codex engagement metrics TODO
