+++ server/index.ts
@@ -498,14 +498,9 @@ async function main(): Promise<void> {
   }
 
   // Periodic rate limit snapshot recording (every 5 minutes)
-  let lastRateLimitSnapshot = 0;
-  const RATE_LIMIT_SNAPSHOT_INTERVAL = 5 * 60 * 1000;
   setInterval(() => {
-    const now = Date.now();
-    if (now - lastRateLimitSnapshot < RATE_LIMIT_SNAPSHOT_INTERVAL) return;
     const account = getAccountTelemetry();
     if (!account || account.fiveHourUsedPercent < 0) return;
-    lastRateLimitSnapshot = now;
     recordRateLimitSnapshot({
       fiveHourPercent: account.fiveHourUsedPercent,
       fiveHourResetsAt: account.fiveHourResetsAt,
@@ -513,7 +508,7 @@ async function main(): Promise<void> {
       sevenDayResetsAt: account.sevenDayResetsAt,
       timestamp: new Date().toISOString(),
     });
-  }, 60_000);
+  }, 5 * 60 * 1000);
 
   // Schedule daily retention cleanup
   setInterval(() => {
